Deleting a Representation is also deleting PREMIS file

+++ roda-core/roda-core/src/main/java/org/roda/core/plugins/plugins/internal/DeleteRODAObjectPlugin.java
@@ -9,6 +9,7 @@ package org.roda.core.plugins.plugins.internal;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -313,14 +314,16 @@ public class DeleteRODAObjectPlugin<T extends IsRODAObject> extends AbstractPlug
           reportItem.addPluginDetails("Could not delete file related incidences: " + e.getMessage());
         }
 
-        // removing PREMIS file
-        try {
-          String pmId = URNUtils.getPremisPrefix(PreservationMetadataType.FILE) + file.getId();
-          model.deletePreservationMetadata(PreservationMetadataType.FILE, file.getAipId(), file.getRepresentationId(),
-            pmId, false);
-        } catch (RequestNotValidException | NotFoundException | GenericException | AuthorizationDeniedException e) {
-          reportItem.addPluginDetails("Could not delete associated PREMIS file: " + e.getMessage());
-        }
+    // removing PREMIS file
+    if (!file.isDirectory()) {
+      try {
+        String pmId = URNUtils.getPremisPrefix(PreservationMetadataType.FILE) + file.getId();
+        model.deletePreservationMetadata(PreservationMetadataType.FILE, file.getAipId(), file.getRepresentationId(),
+          pmId, file.getPath(), false);
+      } catch (RequestNotValidException | NotFoundException | GenericException | AuthorizationDeniedException e) {
+        reportItem.addPluginDetails("Could not delete associated PREMIS file: " + e.getMessage());
+      }
+    }
 
         report.addReport(reportItem.setPluginState(state));
         PluginHelper.updatePartialJobReport(this, model, reportItem, true, job);
@@ -383,14 +386,14 @@ public class DeleteRODAObjectPlugin<T extends IsRODAObject> extends AbstractPlug
           reportItem.addPluginDetails("Could not delete representation related incidences: " + e.getMessage());
         }
 
-        // removing PREMIS file
-        try {
-          String pmId = URNUtils.getPremisPrefix(PreservationMetadataType.REPRESENTATION) + representation.getId();
-          model.deletePreservationMetadata(PreservationMetadataType.REPRESENTATION, representation.getAipId(),
-            representation.getId(), pmId, false);
-        } catch (RequestNotValidException | NotFoundException | GenericException | AuthorizationDeniedException e) {
-          reportItem.addPluginDetails("Could not delete associated PREMIS file: " + e.getMessage());
-        }
+    // removing PREMIS file
+    try {
+      String pmId = URNUtils.getPremisPrefix(PreservationMetadataType.REPRESENTATION) + representation.getId();
+      model.deletePreservationMetadata(PreservationMetadataType.REPRESENTATION, representation.getAipId(),
+        representation.getId(), pmId, Collections.EMPTY_LIST, false);
+    } catch (RequestNotValidException | NotFoundException | GenericException | AuthorizationDeniedException e) {
+      reportItem.addPluginDetails("Could not delete associated PREMIS file: " + e.getMessage());
+    }
 
         report.addReport(reportItem.setPluginState(state));
         PluginHelper.updatePartialJobReport(this, model, reportItem, true, job);
